ECA-2278: fixed yet again.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@12737 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/ra/raadmin/EndEntityProfileSessionBean.java
@@ -196,8 +196,8 @@ public class EndEntityProfileSessionBean implements EndEntityProfileSessionLocal
     public int findFreeEndEntityProfileId() {
         final ProfileID.DB db = new ProfileID.DB() {
             @Override
-            public boolean isFree(Integer i) {
-                return EndEntityProfileData.findById(EndEntityProfileSessionBean.this.entityManager, i.intValue())==null;
+            public boolean isFree(int i) {
+                return EndEntityProfileData.findById(EndEntityProfileSessionBean.this.entityManager, i)==null;
             }
         };
         return ProfileID.getNotUsedID(db);
